requirements.txt
django
djangorestframework
openai
openpyxl
fpdf
google-cloud-secret-manager
python-docx
boto3
